refactor(worker): extract HashRing into its own module, decouple from Worker trait
Finding 5 from .claude/plans/2026-04-14-worker-module-followup-cleanup.md.
Pulls the 118-line consistent-hash-ring data structure out of the
2134-line registry.rs into a focused worker/hash_ring.rs, and narrows
its public API so the type no longer depends on the Worker trait.

What changed
- Add model_gateway/src/worker/hash_ring.rs (new file, ~190 lines incl.
  tests). Contains VIRTUAL_NODES_PER_WORKER, `HashRing`, `hash_position`,
  and six unit tests that cover: empty ring, ring length with virtual
  nodes, deterministic lookup, unhealthy-worker skip, all-unhealthy
  fallback, and owned-String iterator input.
- model_gateway/src/worker/registry.rs:
  * Delete the 118-line HashRing block (struct + impl + const).
  * Update the file-header doc comment to reference the new module.
  * Drop `use std::collections::HashSet` from the HashRing usage (kept
    for other uses in the file).
  * Import `hash_ring::HashRing` from the sibling module.
  * Update `rebuild_hash_ring` to construct the ring from a URL
    iterator: `HashRing::new(workers.value().iter().map(|w| w.url()))`
    instead of `HashRing::new(&workers)`.
- model_gateway/src/worker/mod.rs:
  * Register `pub mod hash_ring;`.
  * Split the re-export — `pub use hash_ring::HashRing;` + `pub use
    registry::WorkerRegistry;` — so downstream callers still write
    `use crate::worker::HashRing` without touching imports.
- model_gateway/src/policies/prefix_hash.rs (5 test sites) and
  model_gateway/src/policies/consistent_hashing.rs (2 test sites):
  * Adapt callers to the new iterator-based signature:
    `HashRing::new(workers.iter().map(|w| w.url()))`.

API change: `HashRing::new`
- Before: `pub fn new(workers: &[Arc<dyn Worker>]) -> Self`
- After:  `pub fn new<I>(urls: I) -> Self where I: IntoIterator, I::Item: AsRef<str>`

  Accepts `&[&str]`, `&[String]`, owned `Vec<String>`, or any iterator
  of string-like items. The rebuild path in `registry.rs` feeds the
  URLs directly from the model index without allocating an intermediate
  Vec, matching the previous zero-alloc hot path.

Why
- HashRing's only dependency on the `Worker` trait was `worker.url()`
  in its constructor. Threading the full trait object through the
  signature coupled a pure consistent-hash data structure to the
  gateway's worker abstraction. Narrowing the constructor to a URL
  iterator removes that coupling so `HashRing` could, in a later PR,
  move to a shared `common/` crate if the gossip/mesh layers ever
  need the same ring.
- `registry.rs` had grown to 2134 lines mixing the ring implementation
  with worker-registry wiring, model index management, and the state
  subscriber. Extracting the ring drops registry.rs by ~115 net lines
  and gives reviewers a focused 190-line file to touch when tuning
  the ring.
- The inline test module on `hash_ring.rs` fills a real coverage gap
  — the ring previously had zero direct tests (the policy tests
  exercised it transitively, but a broken ring would surface in
  policy-level assertions, not as a localized failure).

Verification
- cargo check -p smg — clean.
- cargo clippy -p smg --all-targets -- -D warnings — clean.
- cargo clippy -p smg-golang --all-targets -- -D warnings — clean.
- cargo check -p smg-python — clean.
- cargo test -p smg --lib — 544 passed; 0 failed; 4 ignored
  (+6 new unit tests for HashRing).

Refs: Finding 5 in .claude/plans/2026-04-14-worker-module-followup-cleanup.md
