+++ model_gateway/src/policies/prefix_hash.rs
@@ -264,7 +264,7 @@ mod tests {
     fn test_prefix_hash_consistent_routing() {
         let policy = PrefixHashPolicy::with_defaults();
         let workers = create_workers(&["http://w1:8000", "http://w2:8000", "http://w3:8000"]);
-        let ring = Arc::new(HashRing::new(&workers));
+        let ring = Arc::new(HashRing::new(workers.iter().map(|w| w.url())));
 
         // Same tokens should always route to same worker
         let tokens: Vec<u32> = vec![1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
@@ -288,7 +288,7 @@ mod tests {
     fn test_different_prefixes_distribute() {
         let policy = PrefixHashPolicy::with_defaults();
         let workers = create_workers(&["http://w1:8000", "http://w2:8000", "http://w3:8000"]);
-        let ring = Arc::new(HashRing::new(&workers));
+        let ring = Arc::new(HashRing::new(workers.iter().map(|w| w.url())));
 
         let mut distribution = std::collections::HashMap::new();
 
@@ -318,7 +318,7 @@ mod tests {
             ..Default::default()
         });
         let workers = create_workers(&["http://w1:8000", "http://w2:8000", "http://w3:8000"]);
-        let ring = Arc::new(HashRing::new(&workers));
+        let ring = Arc::new(HashRing::new(workers.iter().map(|w| w.url())));
 
         // Two sequences with same first 5 tokens should route to same worker
         let tokens1: Vec<u32> = vec![1, 2, 3, 4, 5, 100, 200, 300];
@@ -345,7 +345,7 @@ mod tests {
     fn test_no_tokens_returns_none() {
         let policy = PrefixHashPolicy::with_defaults();
         let workers = create_workers(&["http://w1:8000"]);
-        let ring = Arc::new(HashRing::new(&workers));
+        let ring = Arc::new(HashRing::new(workers.iter().map(|w| w.url())));
 
         // Empty tokens
         let tokens: Vec<u32> = vec![];
@@ -377,7 +377,7 @@ mod tests {
         let workers = create_workers(&["http://w1:8000"]);
         workers[0].set_status(WorkerStatus::NotReady);
 
-        let ring = Arc::new(HashRing::new(&workers));
+        let ring = Arc::new(HashRing::new(workers.iter().map(|w| w.url())));
         let tokens: Vec<u32> = vec![1, 2, 3];
         let info = SelectWorkerInfo {
             tokens: Some(&tokens),
